fix: kcms/input and kcms/touchpad should not be built if XInput is not present.
Previously the presence of evdev headers would enable the kcms/input KCM module, and the presence of synaptics driver headers would enable the touchpad module.
However, both kcms unconditionally depend on XInput in some way in a (sub) CMakelists.txt file and deliberately break the build if it is not present or not properly detected.

As per the convention in the kcm modules shipped by plasma-desktop, inclusion of the kcms/input and kcms/touchpad subdirectory is made conditional on XInput being present and properly detected during the cmake configure stage.